% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletters.R
\name{create_palette}
\alias{create_palette}
\title{Create a Color Palette of n Colors from an Image}
\usage{
create_palette(img, n, prominent.ord = FALSE, hue = NA, ...)
}
\arguments{
\item{img}{Path to image file.}

\item{n}{`n` distinct colors.}

\item{prominent.ord}{Option to choose prominent colors instead of segmenting
the colors from a random starting color.}

\item{...}{Additional arguments passed to \link[cuttlefish]{find_segmented}.}
}
\value{
Vector of distinct hex colors.
}
\description{
Creates a color palette of n distinct colors from a provided image.
}
\examples{
  # return 20 semi-evenly distributed colors from some image
  create_palette('path/to/image.png', n = 20)

  # return the 10 most prominent colors from some image
  create_palette('path/to/image.jpg', n = 10, prominent.order = T)
}
