% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorer.R
\name{distinct_hsv}
\alias{distinct_hsv}
\title{Find the n Most Distinct Colors}
\usage{
distinct_hsv(x, n)
}
\arguments{
\item{x}{Matrix of RGB colors.}

\item{n}{`n` distinct colors to be output.}
}
\description{
`distint_hsv` selects the most distinct set of colors from a set of colors.
}
