% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinctions.R
\name{euclid_distance}
\alias{euclid_distance}
\title{Calculate the Distance between a Point and a Vector}
\usage{
euclid_distance(x, y)
}
\arguments{
\item{x}{A numerical value.}

\item{y}{A numerical value or vector.}
}
\value{
A vector of distances
}
\description{
Takes a point `x` and calculates the euclidean distance from a
vector `y`.
}
