% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorer.R
\name{extract_colors}
\alias{extract_colors}
\title{Find Unique RGB Colors in an Image}
\usage{
extract_colors(image)
}
\arguments{
\item{image}{The image to load.}
}
\value{
Returns a vector of hex colors in the image.
}
\description{
`extract_colors` takes in an image and returns all of the
unique colors in the image in RGB color-space.
}
