% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorer.R
\name{find_prominent}
\alias{find_prominent}
\title{Create a Palette of the Most Prominent Colors of an Image}
\usage{
find_prominent(hexes, n)
}
\arguments{
\item{hexes}{A vector of hex colors.}

\item{n}{top `n` most prominent colors.}
}
\value{
Vector of the most prominent hex colors.
}
\description{
Creates a color palette from the most prominent `n` colors of a
    provided image.
}
