% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorer.R
\name{find_segmented}
\alias{find_segmented}
\title{Find the n Most Distinct Colors}
\usage{
find_segmented(hexes, n)
}
\arguments{
\item{hexes}{A vector of hex colors.}

\item{n}{`n` distinct colors to be output.}
}
\value{
Matrix of distinct hsv colors.
}
\description{
`find_segmented` selects the most distinct set of colors from a set of colors.
}
