% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorer.R
\name{prominent_palette}
\alias{prominent_palette}
\title{Create a Palette of the Most Prominent Colors of an Image}
\usage{
prominent_palette(img, n)
}
\arguments{
\item{img}{Path to image file.}

\item{n}{top `n` most prominent colors.}
}
\value{
Vector of the most prominent hex colors.
}
\description{
Creates a color palette from the most prominent `n` colors of a 
    provided image.
}
