% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huers.R
\name{select_hue}
\alias{select_hue}
\title{Subset hex colors by specific hues}
\usage{
select_hue(hexes, hue)
}
\arguments{
\item{hexes}{A vector of hex colors.}

\item{hue}{The hue of colors to be subset. Hues include 'red', 'yellow',
'green', 'cyan', 'blue' and 'magenta'.}
}
\value{
A vector of hex colors belonging to the selected hue.
}
\description{
Finds the colors belonging to a specific hue.
}
