% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletters.R
\name{create_palette}
\alias{create_palette}
\title{Create a Color Palette of n Colors from an Image}
\usage{
create_palette(img, n, prominent.ord = FALSE, ...)
}
\arguments{
\item{img}{Path to image file.}

\item{n}{`n` distinct colors.}

\item{prominent.ord}{Option to choose prominent colors instead of segmenting
the colors from a random starting color.}

\item{...}{Additional arguments passed to \link[cuttlefish]{find_segmented}.}
}
\value{
Vector of distinct hex colors.
}
\description{
Creates a color palette of n distinct colors from a provided image.
}
